Update "getting-started" verification step
I believe the command for verifying the dapr installation is missing the `--help` flag. Running the `dapr` command on my machine (arm64 macOS) gives no output and led me to attempt to reinstall the tool multiple times before realizing the oversight. Running it with the `--help` flag resolved the issue for me. I don't have ready access to a Windows command prompt so I can't test if this is platform specific.

Signed-off-by: Andrew Todd <126006840+aqtodd-abnormal@users.noreply.github.com>